implemented the one-file one purpose
